Checkpoint before ui-decision.
This commit is a checkpoint before I decide what to do about Ui's that
depend on the capabilities.

In particular, 'DocumentContainer caps' is troublesome, because in the
UI-instances for it, which should run the designated uis for the
values contained inside it, we're unable to unify the 'caps' of the
'ui' method to the 'caps' of 'a ~ DocumentContainer caps', because they are
'stuck' in the type variable 'a' of the 'Ui' instance.